Data inputs for a three-species Georges Bank CEATTLE model

A data list containing inputs for a multispecies CEATTLE model fit to three groundfish species on Georges Bank. See BS2017SS for format details.

Usage

GeorgesBank3spp

Format

A list with the same structure as BS2017SS.