Make a vector of colors.

A subset of rich.colors by Arni Magnusson from the gplots package, with the addition of alpha transparency (which is now available in the gplots version as well)

Usage

rich.colors.short(n, alpha = 1)

Arguments

n

Number of colors to generate.

alpha

Alpha transparency value for all colors in vector. Value is passed to rgb function.

Author

Arni Magnusson, Ian Taylor