Management strategy evaluation performance metric summary

Management strategy evaluation performance metric summary

Usage

mse_summary(mse, om_only = FALSE)

Arguments

mse

MSE runs from run_mse or load_mse

om_only

only include performance metrics from OMs

Value

Alist of two data.frames with MSE summary statistics of performance metrics including: data.frame 1

Average annual catch across projection years and simulations per fleet and across fleets

Average interannual variation in catch (IAV) across projection years (n) per fleet and across fleets

% of years in which the fishery is closed across simulations (s)

Average relative mean squared error in estimate of spawning biomass in the terminal year across simulations

% of years in which the population is perceived as undergoing overfishing as determined from F_Limit across simulations via build_hcr in the EM

% of years in which the population is perceived to be overfished as determined from B_Limit across simulations via build_hcr in the EM

% of years in which the population is undergoing overfishing as determined from the “true” F_Limit across simulations via build_hcr in the OM

% of years in which the population is overfished as determined from the “true” B_Limit across simulations via build_hcr in the OM

Average ratio of spawning biomass over B_target in the terminal year across simulations in the OM 10-14. Terminal biomass, SSB, SSB depletion (relative to equilibrium), SSB depletion (relative to dynamic SB0)